+++ RELEASES.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## [v0.13.8](https://github.com/ava-labs/coreth/releases/tag/v0.13.8)
+- Update geth dependency to v1.13.14
+- eupgrade: lowering the base fee to 1 nAVAX
+- eupgrade/cancun: verify no blobs in header
+- Supports ACP-118 message types
+- Gets network upgrade timestamps from avalanchego
+- Remove cross-chain handlers
+
 ## [v0.13.7](https://github.com/ava-labs/coreth/releases/tag/v0.13.7)
 - Add EUpgrade base definitions
 - Remove Block Status
@@ -14,7 +22,7 @@
 
 ## [v0.13.5](https://github.com/ava-labs/coreth/releases/tag/v0.13.5)
 - Bump AvalancheGo to v1.11.7
-- Bump golang version requirement to 1.21.11
+- Bump golang version requirement to 1.21.12
 - Switches timestamp log back to "timestamp" (as was before v0.13.4)
 - Add missing fields to "toCallArg"
 - Fix state sync ETA overflow